fix: add watch txn unit test
