+++ common/meta/api/src/kv_api.rs
@@ -35,6 +35,17 @@ pub trait ApiBuilder<T>: Clone {
     async fn build_cluster(&self) -> Vec<T>;
 }
 
+fn prefix_of_string(s: &str) -> String {
+    let l = s.len();
+    let a = s.chars().nth(l - 1).unwrap();
+    return format!("{}{}", s[..l - 2].to_string(), (a as u8 + 1) as char);
+}
+
+// return watch prefix (start, end) tuple
+pub fn get_start_and_end_of_prefix(prefix: &str) -> (String, String) {
+    return (prefix.to_string(), prefix_of_string(prefix));
+}
+
 #[async_trait]
 pub trait KVApi: Send + Sync {
     async fn upsert_kv(&self, req: UpsertKVReq) -> Result<UpsertKVReply, MetaError>;
